safari-600.7-branch/LayoutTests/fast/css/viewport-vmax-border.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script>
            function runTest() {
                description('Tests that vmax unit works for border-width property');
				
                shouldBeTrue("document.getElementById('borderDiv').offsetHeight > 100");
                isSuccessfullyParsed();
            }
        </script>
        <style>
            #borderDiv{
                width: 100px;
                height: 100px;
                background: green;
                border: 5vmax solid red;
            }
        </style>
    </head>
    <body onload="runTest()">
        <div id="borderDiv"></div>
    </body>
</html>
