Safari-609.1.11.1/LayoutTests/fast/css/viewport-vmax-border-expected.txt

Tests that vmax unit works for border-width property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('borderDiv').offsetHeight > 100 is true
PASS successfullyParsed is true

TEST COMPLETE
